% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_available_countries.R
\name{ph_available_countries}
\alias{ph_available_countries}
\title{Get Available Countries}
\usage{
ph_available_countries()
}
\value{
A tibble containing the following columns:
\describe{
\item{countryCode}{The ISO 3166-1 alpha-2 country code (e.g., "DE" for Germany).}
\item{name}{The name of the country (e.g., "Germany").}
}
}
\description{
This function retrieves a list of available countries for which public holiday data is provided by the Nager.Date API.
}
\examples{
# Get the list of available countries
countries <- ph_available_countries()
print(countries)

}
\references{
Nager.Date API documentation: \url{https://date.nager.at/Api}
}
\seealso{
\code{\link{ph_public_holidays}} for retrieving public holidays for a specific country and year.
}
