% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_country_info.R
\name{ph_country_info}
\alias{ph_country_info}
\title{Get Country Information}
\usage{
ph_country_info(country_code)
}
\arguments{
\item{country_code}{A character string representing the ISO 3166-1 alpha-2 country code (e.g., "DE" for Germany).}
}
\value{
A tibble containing the following columns:
\describe{
\item{commonName}{The common name of the country (e.g., "Germany").}
\item{officialName}{The official name of the country (e.g., "Federal Republic of Germany").}
\item{countryCode}{The ISO 3166-1 alpha-2 country code (e.g., "DE").}
\item{region}{The region where the country is located (e.g., "Europe").}
\item{borders}{A list of country codes for countries that share a border with the specified country.}
}
}
\description{
This function retrieves detailed information about a specific country using the Nager.Date API.
}
\examples{
# Get information about Germany
country_info <- ph_country_info("DE")
print(country_info)

}
\references{
Nager.Date API documentation: \url{https://date.nager.at/Api}
}
\seealso{
\code{\link{ph_available_countries}} for retrieving a list of available countries.
\code{\link{ph_public_holidays}} for retrieving public holidays for a specific country and year.
}
