% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_long_weekends.R
\name{ph_long_weekends}
\alias{ph_long_weekends}
\title{Get Long Weekends for a Country in a Given Year}
\usage{
ph_long_weekends(country_code, year)
}
\arguments{
\item{country_code}{A character string representing the ISO 3166-1 alpha-2 country code (e.g., "DE" for Germany).}

\item{year}{An integer representing the year (e.g., 2025).}
}
\value{
A tibble containing the following columns:
\describe{
\item{startDate}{The start date of the long weekend (format: "YYYY-MM-DD").}
\item{endDate}{The end date of the long weekend (format: "YYYY-MM-DD").}
\item{dayCount}{The number of days in the long weekend.}
\item{needBridgeDay}{A logical value indicating whether a bridge day is needed to create the long weekend.}
\item{bridgeDays}{A list of bridge days (if any) required to create the long weekend.
}
}
\description{
This function retrieves a list of long weekends (extended weekends due to public holidays) for a specific country and year using the Nager.Date API.
}
\examples{
# Get long weekends for Germany in 2025
long_weekends <- ph_long_weekends("DE", 2025)
print(long_weekends)

}
\references{
Nager.Date API documentation: \url{https://date.nager.at/Api}
}
\seealso{
\code{\link{ph_public_holidays}} for retrieving public holidays for a specific country and year.
\code{\link{ph_available_countries}} for retrieving a list of available countries.
}
